requirements.txt
aiohttp~=3.9.5
aiosignal~=1.3.1
annotated-types~=0.6.0
anyio~=4.3.0
APScheduler~=3.10.4
attrs~=23.2.0
certifi~=2024.2.2
charset-normalizer~=3.3.2
colorama~=0.4.6
configparser~=7.0.0
distro~=1.9.0
feedparser~=6.0.11
frozenlist~=1.4.1
gitdb~=4.0.11
GitPython~=3.1.43
greenlet~=3.0.3
h11~=0.14.0
html2text~=2024.2.26
httpcore~=1.0.5
httpx~=0.27.0
idna~=3.7
khl.py~=0.3.17
multidict~=6.0.5
openai~=1.28.0
ping3~=4.0.8
pycryptodomex~=3.20.0
pydantic~=2.7.1
pydantic-settings~=2.2.1
pydantic_core~=2.18.2
pyperclip~=1.8.2
python-dateutil~=2.9.0.post0
python-dotenv~=1.0.1
pytz~=2024.1
requests~=2.31.0
riddlesolver~=0.1.20
sgmllib3k~=1.0.0
six~=1.16.0
smmap~=5.0.1
sniffio~=1.3.1
SQLAlchemy~=2.0.29
tqdm~=4.66.4
typing_extensions~=4.11.0
tzdata~=2024.1
tzlocal~=5.2
urllib3~=2.2.1
yarl~=1.9.4
